+++ .flake8
@@ -10,4 +10,4 @@ exclude =
     .venv/**
 
 per-file-ignores =
-    **/**/__init__.py:F401
+    **/__init__.py:F401
